refactor: factor LauncherConfig out from vite-dev-overrides
